Move top-level CLI to lib
